Перламутровые пуговицы. Вторая проверка
