+++ js/create-previews.js
@@ -1,7 +1,3 @@
-import { initBigPicture } from './big-picture.js';
-import { getData } from './get-data-api.js';
-import { debounce } from './utils/debounce.js';
-
 const similarImagesBlock = document.querySelector('.pictures');
 const similarImagesTemplate = document.querySelector('#picture').content.querySelector('.picture');
 const imgFilters = document.querySelector('.img-filters');
@@ -30,57 +26,31 @@ const createPreviews = (images) => {
   imgFilters.classList.remove('img-filters--inactive');
 };
 
-const initPressButtonDefaultImages = debounce(() => {
-  defaultButton.classList.add('img-filters__button--active');
-  randomButton.classList.remove('img-filters__button--active');
-  discussedButton.classList.remove('img-filters__button--active');
-  getData((images) => {
-    createPreviews(images);
-    initBigPicture(images);
+const showDefaultImages = (cb) => {
+  defaultButton.addEventListener('click', () => {
+    defaultButton.classList.add('img-filters__button--active');
+    randomButton.classList.remove('img-filters__button--active');
+    discussedButton.classList.remove('img-filters__button--active');
+    cb();
   });
-});
-
-const initPressButtonRandomImages = debounce(() => {
-  defaultButton.classList.remove('img-filters__button--active');
-  randomButton.classList.add('img-filters__button--active');
-  discussedButton.classList.remove('img-filters__button--active');
-  getData((images) => {
-    const randomImages = images.slice();
-
-    const shuffleArrayImages = (ArrayImages) => {
-      let j, temp;
-      for (let i = ArrayImages.length - 1; i > 0; i--) {
-        j = Math.floor(Math.random()*(i + 1));
-        temp = ArrayImages[j];
-        ArrayImages[j] = ArrayImages[i];
-        ArrayImages[i] = temp;
-      }
-      return ArrayImages;
-    };
-
-    const randomArrayImages = shuffleArrayImages(randomImages).slice(0, 10);
-    createPreviews(randomArrayImages);
-    initBigPicture(randomArrayImages);
-  });
-});
-
-const initPressButtonDiscussedImages = debounce(() => {
-  defaultButton.classList.remove('img-filters__button--active');
-  randomButton.classList.remove('img-filters__button--active');
-  discussedButton.classList.add('img-filters__button--active');
-  const compareCommentsImages = (imageA, imageB) => imageB.comments.length - imageA.comments.length;
-  getData((images) => {
-    const discussedImages = images.slice();
-    discussedImages.sort(compareCommentsImages);
-    createPreviews(discussedImages);
-    initBigPicture(discussedImages);
-  });
-});
-
-const initPressButtonImagesFilters = () => {
-  defaultButton.addEventListener('click', initPressButtonDefaultImages);
-  randomButton.addEventListener('click', initPressButtonRandomImages);
-  discussedButton.addEventListener('click', initPressButtonDiscussedImages);
 };
 
-export { createPreviews, initPressButtonImagesFilters };
+const showRandomImages = (cb) => {
+  randomButton.addEventListener('click', () => {
+    defaultButton.classList.remove('img-filters__button--active');
+    randomButton.classList.add('img-filters__button--active');
+    discussedButton.classList.remove('img-filters__button--active');
+    cb();
+  });
+};
+
+const showDiscussedImages = (cb) => {
+  discussedButton.addEventListener('click', () => {
+    defaultButton.classList.remove('img-filters__button--active');
+    randomButton.classList.remove('img-filters__button--active');
+    discussedButton.classList.add('img-filters__button--active');
+    cb();
+  });
+};
+
+export { createPreviews, showDefaultImages, showRandomImages, showDiscussedImages };
